feat: Phase 1 Foundation — Decimal types, Ruff, mypy, dependency updates
- 1.1: Float → Decimal(Numeric) for weight_kg, declared_value_usd,
  delivery_cost_rub; Alembic migration; remove float() conversions
- 1.2: Widen dependency constraints (fastapi, uvicorn, aiomysql, httpx
  <1.0.0); upgrade redis 4.x → 5+ (<8.0.0); MySQL 8.0 → 8.4
- 1.3: Replace black + isort with Ruff for linting and formatting
- 1.4: Add mypy with pydantic plugin; fix type errors across codebase
- 1.5: Extract magic numbers to Settings (delivery formula coefficients,
  batch size, lock TTL, scheduler interval, cache TTLs)
- 1.6: Fix pyproject.toml PEP 508 syntax, add project description,
  consolidate pytest.ini into pyproject.toml

+++ app/models/parcel.py
@@ -1,6 +1,7 @@
+from decimal import Decimal
 from uuid import uuid4
 
-from sqlalchemy import Float, ForeignKey, String
+from sqlalchemy import ForeignKey, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.db.base import Base
@@ -31,14 +32,14 @@ class Parcel(Base):
         doc="Human-readable parcel name (e.g. 'Laptop bag').",
     )
 
-    weight_kg: Mapped[float] = mapped_column(
-        Float,
+    weight_kg: Mapped[Decimal] = mapped_column(
+        Numeric(10, 3),
         nullable=False,
         doc="Weight in kilograms, used for cost calculation.",
     )
 
-    declared_value_usd: Mapped[float] = mapped_column(
-        Float,
+    declared_value_usd: Mapped[Decimal] = mapped_column(
+        Numeric(12, 2),
         nullable=False,
         doc="Declared customs value in USD.",
     )
@@ -49,8 +50,8 @@ class Parcel(Base):
         doc="Anonymous session ID that owns this parcel.",
     )
 
-    delivery_cost_rub: Mapped[float | None] = mapped_column(
-        Float,
+    delivery_cost_rub: Mapped[Decimal | None] = mapped_column(
+        Numeric(12, 2),
         nullable=True,
         doc="Final delivery cost in RUB, calculated asynchronously.",
     )
